requirements.txt
# Core LLM and Analysis Libraries
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
openai>=1.10.0
anthropic>=0.18.0

# Alternative: LlamaIndex (uncomment if you prefer)
# llama-index>=0.9.0
# llama-index-llms-openai>=0.1.0

# Code Parsing and Analysis
tree-sitter>=0.20.4
tree-sitter-java>=0.20.2
javalang>=0.13.0
lizard>=1.17.10  # Code complexity analysis

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.2  # Token counting
pydantic>=2.0.0  # Data validation
rich>=13.0.0  # Beautiful CLI output

# File handling
pathspec>=0.11.0  # .gitignore handling


